Some minor updates
Nice job. I added some additional comments. In general, class comments should describe a thing, which you did. And method comments describe an action, which you did. It is considered good practice to start with the action in the method comment.
It is also good to put file headers with the code. It gives you context.